+++ assignment_tving/assignment_tving/ViewController.swift
@@ -1,0 +1,19 @@
+//
+//  ViewController.swift
+//  assignment_tving
+//
+//  Created by Minjoo Kim on 2023/04/09.
+//
+
+import UIKit
+
+class ViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        // Do any additional setup after loading the view.
+    }
+
+
+}
+
